requirements.txt
fastapi
uvicorn
requests
yt_dlp
python-dotenv